LayoutTests/imported/w3c/web-platform-tests/scroll-animations/animation-trigger/timeline-trigger-source-starts-in-trigger-range.tentative-expected.txt

Target
Source

FAIL Already in-view source triggers animation assert_equals: expected "200px" but got "104.59375px"
